Complete Applications/Document Intelligence Extract Important Info from Documents/requirements.txt
altair==4.2.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
attrs==21.4.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.0
blinker==1.4
blis==0.7.7
cachetools==5.0.0
catalogue==2.0.7
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
ci-info==0.2.0
click @ file:///C:/ci/click_1646038595831/work
colorama @ file:///tmp/build/80754af9/colorama_1607707115595/work
configobj==5.0.6
configparser==5.2.0
cymem==2.0.6
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
docopt==0.6.2
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.2.0/en_core_web_sm-3.2.0-py3-none-any.whl
entrypoints==0.4
etelemetry==0.3.0
executing==0.8.3
fastjsonschema==2.15.3
filelock==3.6.0
fitz==0.0.1.dev2
gitdb==4.0.9
GitPython==3.1.27
gTTS==2.2.4
httplib2==0.20.4
idna==3.3
importlib-metadata==4.11.3
ipykernel==6.13.0
ipython==8.2.0
ipython-genutils==0.2.0
ipywidgets==7.7.0
isodate==0.6.1
jedi==0.18.1
jellyfish==0.9.0
Jinja2==3.1.1
joblib @ file:///tmp/build/80754af9/joblib_1635411271373/work
jsonschema==4.4.0
jupyter-client==7.3.0
jupyter-core==4.10.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.0
langcodes==3.3.0
lxml==4.8.0
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
mistune==0.8.4
murmurhash==1.0.7
nbclient==0.6.0
nbconvert==6.5.0
nbformat==5.3.0
nest-asyncio==1.5.5
networkx==2.8
nibabel==3.2.2
nipype==1.7.1
nltk @ file:///opt/conda/conda-bld/nltk_1645628263994/work
notebook==6.4.11
numpy==1.22.3
packaging==21.3
pandas==1.4.2
pandocfilters==1.5.0
parso==0.8.3
pathy==0.6.1
pickleshare==0.7.5
Pillow==9.1.0
preshed==3.0.6
prometheus-client==0.14.1
prompt-toolkit==3.0.29
protobuf==3.20.1
prov==2.0.0
psutil==5.9.0
pure-eval==0.2.2
pyarrow==7.0.0
pycparser==2.21
pydantic==1.8.2
pydeck==0.7.1
pydot==1.4.2
Pygments==2.12.0
Pympler==1.0.1
PyMuPDF==1.19.6
pyparsing==3.0.8
pyrsistent==0.18.1
pytesseract==0.3.9
python-dateutil==2.8.2
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
pywin32==303
pywinpty==2.0.5
pyxnat==1.3
pyzmq==22.3.0
rdflib==6.1.1
regex==2022.4.24
requests==2.27.1
scikit-learn==1.0.2
scipy==1.8.0
segtok==1.5.11
semver==2.13.0
Send2Trash==1.8.0
simplejson==3.17.6
six==1.16.0
sklearn==0.0
smart-open==5.2.1
smmap==5.0.0
soupsieve==2.3.2.post1
spacy-legacy==3.0.9
spacy-loggers==1.0.2
srsly==2.4.3
stack-data==0.2.0
streamlit==1.8.1
tabulate==0.8.9
terminado==0.13.3
thinc==8.0.15
threadpoolctl==3.1.0
tinycss2==1.1.1
toml==0.10.2
toolz==0.11.2
tornado==6.1
tqdm @ file:///C:/ci/tqdm_1650636210717/work
traitlets==5.1.1
traits==6.3.2
typer==0.4.1
typing_extensions==4.2.0
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.9
validators==0.18.2
wasabi==0.9.1
watchdog==2.1.7
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.6.0
wincertstore==0.2
yake==0.4.8
yarg==0.1.9
zipp==3.8.0
